remove numpy int and float

+++ src/classes/field_of_view.py
@@ -1288,7 +1288,7 @@ class Field_of_View():
                     # ids
                     if 'ids' not in _grp:
                         _ids = _grp.create_dataset('ids', (len(_dict['ids']),), dtype='i', data=_dict['ids'])
-                        _ids = np.array(_dict['ids'], dtype=np.int) # save ids
+                        _ids = np.array(_dict['ids'], dtype=np.int32) # save ids
                         _data_attrs.append('ids')
                     elif len(_dict['ids']) != len(_grp['ids']):
                         _change_size_flag.append('id')
@@ -1430,11 +1430,11 @@ class Field_of_View():
                                         self.shared_parameters['allowed_data_types'])
             _region_ids = _type_dic[_data_type]
         else:  
-            if isinstance(_region_ids, int) or isinstance(_region_ids, np.int):
+            if isinstance(_region_ids, int) or isinstance(_region_ids, np.int32):
                 _region_ids = [_region_ids]   
             elif not isinstance(_region_ids, list) and not isinstance(_region_ids, np.ndarray):
                 raise TypeError(f"Wrong input type for region_ids:{_region_ids}")
-            _region_ids = np.array([int(_i) for _i in _region_ids],dtype=np.int)
+            _region_ids = np.array([int(_i) for _i in _region_ids],dtype=np.int32)
         # print
         if _verbose:
             _check_time = time.time()
@@ -1799,7 +1799,7 @@ class Field_of_View():
                 # choose dapi folder
                 if len(_dapi_fds) == 1 or (len(_dapi_fds) > 1 and _dapi_id is None): 
                     _dapi_fd = _dapi_fds[0]
-                elif isinstance(_dapi_id, int) or isinstance(_dapi_id, np.int):
+                elif isinstance(_dapi_id, int) or isinstance(_dapi_id, np.int32):
                     _dapi_fd = _dapi_fds[_dapi_id]
                 else:
                     raise TypeError(f"Wrong input type: {type(_dapi_id)} for _dapi_id.")
@@ -1872,7 +1872,7 @@ class Field_of_View():
         if 'correct_fov_image' not in locals():
             from ..io_tools.load import correct_fov_image
 
-        if isinstance(_bead_id, int) or isinstance(_bead_id, np.int):
+        if isinstance(_bead_id, int) or isinstance(_bead_id, np.int32):
             _ind = int(_bead_id)
         elif isinstance(_bead_id, str):
             for _i, _fd in enumerate(self.annotated_folders):
